refactor: improve token validation and UI structure
